Fix broken B4SOI SHMOD=1 handling
An error in Xyceification of B4SOI introduced a bug in the handling of
self-heating (SHMOD=1).  Instead of DevTemp always being ambient
temperature plus the self-heating temperature rise, the rise was being
accumulated into DevTemp over the entire run, causing horrific
convergence failures.

This commit fixes the error by re-initializing DevTemp to ambient
temperature on each iteration, so that the self-heating rise is
added to it anew instead of accumulating.

It also removes the notes in the release notes and referernce guide
that SHMOD=1 has convergence issues that are under investigation.

Xyce/Xyce#7