Add Initial Support for .MEASURE for .NOISE Analyses
This commit adds the initial support for .MEASURE for .NOISE analyses.
Since .AC and .NOISE are different analysis objects in Xyce, this
new mode is requested with .MEASURE NOISE lines.  This initial commit
supports the AVG, ERR, ERR1, ERR2, EQN/PARAM, INTEG, MAX, MIN, PP
and RMS measures.  Support for the INOISE, ONOISE, DNI() and DNO()
operators is included.

One "hack" done for this feature is because (per SON Bug 1013) the
-r and -a output for .NOISE uses the outputManagerAdapter_.outputAC()
function.  The existing version of that function also updated the AC
mode measures. A new outputManagerAdapter_.outputACwoMeasureUpdates()
function was added that does not update the AC mode measures.  Once
SON Bug 1013 is addressed, this approach might be re-thought.

This commit also updated some functions in the Base object to make
it clear that they support AC, DC and NOISE modes now.

SON Bug 1301
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1301