Move BSIM-SOI 4.6.1 code to new directory structure
Jason (@jcverle) would like to rearrange how we keep Verilog-A models
stored in the repo.  While I'm not quite ready to rearrange the entire
tree, I'm getting ready to implement BSIM-SOI 4.5.0, and am
rearranging the way I store BSIM-SOI 4.x code to be consistent with
his new scheme (which I like and will eventually implement in full, as
the current "examples" directory tree is an artifact of very ancient
ADMS development and not well thought out).

In doing this, I'm removing ALL BSIM-SOI documentation PDFs and
"benchmark_test" netlists, and retaining only the Verilog-A source
code, the README_Xyce file, and the Makefile.

Xyce/Xyce#3