Add BSIM-SOI 4.5.0 as level 70450 MOSFET
It is becoming apparent that BSIM-SOI 4.6.1 is not a suitable
substitute in PDKs that expect BSIM-SOI 4.5.0 or earlier, as there
appears to be a bug in the 4.6.0/4.6.1 model relating to gate
capacitances.

This commit adds the 4.5.0 model as the level=70450 MOSFET.

A rethinking of model levels and their correspondence to BSIM-SOI
versions may be required.

Xyce/Xyce#10