Fix for applying setParam to dependent params
This commit is intended to address this use case and a few others:

.global_param vdi=0.5
V1 1 0 {vdi}
.DC V1 0.1 0.1 1

In the above use case, the device entity parameter V1:DCV0 is a
"dependent parameter", in that it depends on the expression {vdi}.  So,
under normal circumstances, V1:DCV0 = whatever {vdi} is.

However, in the DC sweep line above, V1:DCV0 is set by the sweep.  So
that is supposed take priority over the global parameter {vdi}.

The problem, in Xyce design, is that the DeviceMgr::setParam call, which
is used by the DC sweep to set sweep values, is called *before* all the
various devices call the function
DeviceEntity::updateDependentParameters, which is where the expression
{vdi} would be evaluated and imposed on V1:DCV0.

So, there has to be a mechanism to know when one should skip the
expression evaluation in updateDependentParameters.

With the old expression library, there was a mechanism, but it was well
hidden and pointlessly confusing.    That mechanism was wrapped up
inside the "changed" logic and also the number n_vars in the "Depend"
class.  I used to think that "changed" was only about making espressions
more efficient, and avoiding re-evaluations that were wasteful.  But, it
is also about letting "setParam" have priority when necessary.

This implicit (hidden) logic can't work very easily with newExpression.
One reason is that the newExpression class gets its values via a "pull"
operation thru the group, whereas the old expression library got them
via a "push" using functions like "set_vars" and "set_sim_time".  These
push/set calls were the calls that would provide a value for "changed"
upon return.  Since I don't make those calls with new expression, I
can't very well use the same "changed" based logic.

So instead, I've added a simple unordered_map to keep track of
parameters that have had "setParam" called on them.  If they have, then
they are excluded from "dependent param" operations.  Problem solved,
and much easier for future developers to understand.