Make exclamation point allowed in node/device name
I had not previously allowed an exclamation point in the names of nodes
and devices.  Now I do.

This was prompted by the POWER/power_unsupported_devices.cir test, which
attempts to get power from the YACC!ACC1 device.

Ironically, this is a test that is supposed to exit with error anyway.
But it was exiting too early with the wrong error.