Add documentation for BSIM-SOI 4.5.0
A decision was made to cherry-pick the late-entry BSIM-SOI 4.5.0 code
off of develop and into the 7.1 release.  I am therefore adding it to
the release notes and adding its tables to the reference guide, with
an explanation of why we are supporting the older model and what it is
about the newer one that may cause disagreement with other simulators.

Xyce/Xyce#10
Xyce/Xyce#9