Implement operator for schedule
SCHEDULE is a capability that allows the user to specify a list of max
time steps to be used at different phases of the simulation.  I had
neglected to implement it.

But, implementing it was not too difficult, as it is similar to table,
but doesn't do any interpolation.  I probably need to debug it a little
more, but I have created one unit test, which passes and the one
regression test that uses this feature, MAX_TIME_STEP/max_time_step now
passes as well.