Add Support for DERIV Measure for .MEASURE NOISE
This commit adds support for the DERIV (and DERIVATIVE) measure
for .MEASURE NOISE.  It also refactors the DerivativeEvaluation
class so that the updateAC(), updateDC and updateTran() functions
all use a common set of functions to process the AT and WHEN
conditions.  The TRAN version of the DERIV measure also supports
the RISE, FALL and CROSS qualifiers, so slightly different code
is used to evaluate the AT and WHEN conditions in updateTran().

SON Bug 1301
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1301