Add Support For Touchstone 1 Input Files for YLIN
This commit adds support for Touchstone 1 formatted input files for the
YLIN device.  A summary of the changes is as follows:

  1) The function Model::readTouchStoneFile() now handles opening and
closing the input file.  It also reads the first line of the input file.
If the first non-blank, non-comment line is a [Version] line then the
file is assumed to be Touchstone 2.  If that first line is an Option
line then the file is assumed to be Touchstone 1.

  2) Model::readTouchStone1File() and Model::readTouchStone2File() then
handle the rest of the processing for those specific formats.

  3) The logic for parsing the Option line and each individual line
of network data is common to both formats.  So, that logic was pulled
out into two new functions.

  4) The function Model::readTouchStone1File() handles column wrapping,
which occurs for three or more ports.  Lines that start a complete line
of network data have an odd number of elements, since they start with
a frequency value.  "Column wrapped" lines, that continue the network
data for a given frequency, then have an even number of elements.
Processing is based on that rule, and uses one line of look-ahead.
The error checking that each complete line of network data must have the
same length provides some minimal error checking against mis-formatted
input.

SON Bug 1288
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1288