Partial fix for bug 970
But 970 is about properly documenting which devices will not work with
frequency domain analysis.  Some work had been done on this already,
before I took over the bug.

Recent energy behind this issue was driven by a
user pointing out that some of the information in the guide,
claiming that B sources and devices derived from them such as E sources
didn't work in frequency domain, was wrong.

I have updated the guides in a few places to correct this.  I have added
text to the HB, B-device and "incompatibilities with other spice
simulators" section of the reference guide.  I have also added text to
the ABM chapter of the user guide.

In all places, I basically say, "B sources work in frequency domain
analysis, as long as they are not time dependent".  I have eliminated
mention of E,H,G, etc sources, as I think they don't have the option to
be time dependent.