Add guidance for AC/HB to the E,F,G,H devices
This is related to bug 970 SON.

I previously added guidance w.r.t. the B-source device, to say that it
would only work with frequency domain analysis if it was not
time-dependent.

In that commit, I excluded E,F,G,H devices, even though they are
secretly B-sources, as I mistakenly thought that they didn't have the
option to be time-dependent.  I don't use these devices very much, and
it turns out I was wrong about that.  Tom pointed out that these devices
do allow the user to specify a "value=expression", and the expression
can support anything that a normal B-source can support.  So,
technically they can be time-dependent as well, even though this use
case is not common.

So, now these devices are included in the no-time-dep guidance.

This will be all I do on this bug for 7.1.  While I think this addresses
the immediate problem, which was incorrect information in the guides, it
isn't sufficient to close the bug.  So I have deferred it.