Fix broken reference guide
The 7.1 reference guide failed to build because of a misuse of
"\alltt" instead of "\texttt" in five recently edited files.

"alltt" is the name of an environment that should be used in a
begin/end construct.  \texttt is the command you use to render the
following block in typewriter font.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=970