Make E and G device table consistent with others
A recent commit added a bunch of almost identical text to E, F, G, H,
and B source tables, but in the case of E and G put it outside the
\begin{Device}/\end{Device} environment, making these two devices
render inconsistently with all the others.

This commit moves the \end{Device} to the same place as the three other
cases, so that the text is indented like all the other paragraphs in
the "Comments" section.