Fix Handling of TS1 Input Files for YLIN Device
This commit fixes the handling of Touchstone 1 input files for the
YLIN device for the case of ISC_FD=true.  (The per-port short-circuit
currents are given in that case.)  The formula for calculating the
number-of-ports, from the first complete line of network data, is
now correct for that case.  The code now also checks that the derived
value of the number of ports is an integer >= 1.

Additional refactoring was done to remove duplicate code, and to
simplify the function signatures of various "helper" functions that
are common to processing both Touchstone 1 and Touchstone 2 formatted
input files.

The final change improved the handling of multiple option lines
(that start with #).  The second one (and subsequent ones) are
silently ignored, per the Touchstone standard.

SON Bug 1288
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1288