Add preliminary support for SDT
SDT is the operator for performing numerical time integration.  I am
still iterating on the design of it.   The main implementation
difficulty issue for SDT (and DDT) is that the operator has to be
informed as to when the time step advances.

It is incorrect to assume that the time has advanced every time the
operator::val aor  operator::dx function is called.   And, the parts of
Xyce that would know when time advances don't necessarily have access to the
current expression.  For example, the N_ANP_Transient class doesn't have a list
of expressions to call to advance the time.

The way the old expression library handled this issue was to check if
the "time" or the "dt" value changed, by comparing floating point
numbers.  I wasn't comfortable with doing this, as it feels really
kludgey and makes the code ugly.  (also depending on floating point
numbers being 100% identical is risky). So, I am not doing that.

For the time being I have added a function to the SDT and DDT operators,
"processSuccessfulTimeStep", which is called from another new function
newExpression::processSuccessfulTimeStep.  This works, and I have
created a handful of parser unit tests that use these functions and
pass.

I have *not* done the work yet to call this function from Xyce, so as of
this writing, using SDT in a netlist won't yet work with newExpression.
For the expressions that use the "mainXyceExpressionGroup" (i.e. the
ones in devices), I think this won't be too hard to set up.  But that
will happen in a later commit.