Improve Handling of Blank Lines in TS Input for YLIN
This commit improves the handling of "blank", comment and extra Option
lines in Touchstone 1 and Touchstone 2 input files for the the YLIN
device.  Leading whitespace is now trimmed for all lines in those
input files.  This makes it easier to skip over comment lines (that
start with the ! character or that have leading whitespace before the
! character) and "blank lines" (which either have zero length or only
whitespace on them).  Once the first Option line is found, all
subsequent Option lines should be ignored.  The logic for determining
if a line is a "blank", comment or extra Option line has been
consolidated into two new helper functions.

SON Bugs 1215 and 1288
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1288