Enable Stokhos for Windows builds
In the cmake-refactor branch, Xyce's build is detecting whether or not
to enable stokhos by looking to see if it's present in Trilinos, and
always building in that support if so.

It's failing all those tests.

We have never before enabled Stokhos on Windows, because it wouldn't
have worked before we updated the compiler from Intel 16 to Intel
19 (or something like that).  In fact, there may be issues with
Stokhos support that is unrelated to the cmake refactor.

This commit makes Xyce build in stokhos by default using the (old)
cmake build.  I'm doing this solely to force it to be tested.  It is
not cleverly trying to check if stokhos is present in Trilinos,
though, it's just assuming it is.

Xyce/Xyce#2