Set NOPACK define for Windows builds.
The Windows builds need to have Xyce_RESTART_NOPACK defined to
avoid using binary files for restart.  This define was left out
of the cmake refactor work.  Also the Xyce_STOKOS_ENABLE define
had a typo in the Xyce_config.h.cmake file so it was not correctly
used when it was defined.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/2