Make version check for stokhos consistent with other uses
In most places in Xyce where Xyce_STOKHOS_ENABLE is checked, it is
checked with "#if Xyce_STOKHOS_ENABLE" for some reason.  This makes
the use of "#ifdef" in N_UTL_Version.C inconsistent, and if
Xyce_STOKHOS_ENABLE is defined as zero then Version is saying stokhos
is enabled, but the rest of the code is not.

Xyce/Xyce#2