Undo "make stokhos default" change
In an earlier commit, I tried to force stokhos "on" for Windows based
on an incorrect assumption that our build of Trilinos on Windows had
been done with Stokhos enabled and we weren't correctly handling it.

That was wrong, we don't have it built that way, but were
inconsistently checking for it in N_UTL_Version compared to where it's
tested elsewhere.

I'm undoing the default.

The fix to N_UTL_Version.C I just made on develop should be merged
into cmake-refactor.  That may fix the issue with failing stokhos
tests (which are failing because stokhos isn't in Trilinos).

Xyce/Xyce#2