Initial implementation of ddt, refinement of sdt
This is all still a work in progress, but I got a version of ddt
running.  It works pretty well on the .print line, but isn't working
correctly in the ABM device (Bsrc) yet.

I've revised the book-keeping for time stepping in new expression
somewhat, relative to my previous commit for SDT.

sdt in the new expression library is now passing all the sdt tests
except for 2.