Fix Handling of FROM and TO for .MEASURE DC
This commit fixes the handling of the FROM and TO qualifiers for
DC measures for the cases where only FROM is given, or only TO
is given, on a .MEASURE DC line.  Previously, Xyce only worked
correctly if both, or neither, was given on a .MEASURE DC line.
This change affects all measure types that are supported for .DC.

SON Bug 1306
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1306