Fix for new expression with MPDE
Several of the MPDE tests have been failing with the new expression
library.  Today I took the time to figure out why and fix it.

The reason had to do with my implementation of the "getSolutionVar"
function in the mainXyceExpressionGroup class.  In that implementation,
I pull values out of the "next" solution vector, but the pointer to that
vector that I use is pulled out of the data store class.

This is fine most of the time, but not for Bsrc devices in the MPDE case.
In the MPDE case, the solution vector in data store is a block vector,
and it has a # of blocks equal to the size of the fast time scale.  So,
if I pull values naively out of that vector, I'm always using the first
block.  This will usually be the wrong block to use.

So, instead, I needed to pull the solution vector pointer out of the
"externData" class in the device pacakge, as it only knows about the
specific block currently in use.   This was an easy change to make, and
it fixed 4 or 5 MPDE tests with new Expression.