Updated to handle global $G nodes
Two xyce regression tests were failing due to this issue.  They were:

Certification_Tests/BUG_1430/bug_1430.............EXITED WITH ERROR[sh] (Time:   1s =   0.07cs +   0vs)
C_T/BUG_1661/globalnode_expr_toplev...............EXITED WITH ERROR[sh] (Time:   0s =   0.07cs +   0vs)

The reason that they were failing was that I wasn't lexing the "$"
character properly.  It wasn't included in the allowed list of
characters for a voltage node name.

In the paarser (N_IO_DistToolBase.C file), it has a process where it
converts nodes that are inside of subcircuits to their fully resolved
names.  So, for example, a name might go from "1" to "X1:1".  In that
section of code, there is an exception to this, if a node starts with
"$G", then it is exempt from this process.

In the above tests, the is a Bsrc inside of a subcircuit that depends on
a global node:

B1 a b V={V($g_1)}

Since my expression lexer wasn't recognizing "$", it interpretted the
above as:

B1 a b V={V(g_1)}

So, in the N_IO_DistToolBase.C file, when it got to subcircuit
resolution, the "$G" exception wasn't triggered, and it converted the
node to X1:G_1.   This was not resolvable as this node did not exist.
Thus Xyce exited with error and the test failed.

This was very easy to fix.  I just had to add "$" to the regex for NODE
in the expression lexer.