Remove configure options for OP_START
The "dcop restart" stuff was removed in commit b0546a557 over a year
ago, but the configure and cmake options to enable it remained.

I spotted a user on Github who maintains a build configuration
repo (has reconfigure scripts, with Xyce and Trilinos as git
submodules) and whose reconfigure script for Xyce is enabling
op_restart explicitly, despite the fact that this option does nothing.
That's probably because he spotted the option in the configure help
and is explicitly enabling everything possible (which includes nonfree
and modspec, both of which will be ignored unless the code is present!).

I am removing the option, since it is a no-op.

http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1151