Fix stokhos config code to check properly
When code was added to configure to add the "--enable-stokhos" option,
no check for the existence of Stokhos headers was added, and an
unnecessary AM_CONDITIONAL was put in (which is used only for
conditional compilation in Makefiles, none of which exists).

This commit makes sure that if a user (say, me) accidentally requests
Stokhos support in a Xyce build, but tries to do so when Trilinos
hasn't been built with Stokhos support, that notification of that fact
comes at configure time, not ten minutes into a build when
compilations fail.