Add support for various weird chars in node names
Earlier this week I modified the NODE macro so that it could also
include the "$" character, to support global nodes in expressions.  I
didn't quite do that support correctly, as that character is a special
character in regular expressions, so it needs to preceeded by a
backslash in order to be a literal '$'.  That is fixed now.

I was prompted to fix it more correctly by another test failure, which
has been the INVALID_CHARS/valid_chars_expressions.cir test.  This test
has many, many examples of node names that include various special
characters.  For example V(a>), V(a%), etc.   I had forgotten that these
types of characters are supported in node names (and device names), but
this is a relatively recent development in Xyce, documented in SON
bug 1034.  The title of that bug was "Issues with processing expressions
that contain arithmetic operators (e.g., + - or *) within an otherwise
legal node or device name"

This wasn't too hard to support in the new expression library.  The
regex used to identify a NODE macro needed further expansion.  As many
of these weird characters are wild cards and other types of special
characeters in regex, they also had to be preceded by a backslash.

Most of this commit was focused on the NODE macro, which in the lexer is
used to define what characters are allowed for voltage node names.
However, I also added a little to the macro associated with device
names.    Specifically, I added '+' and '-' to allowe characters for
device names.

Doing this forced me to create a new macro, DEV.  Previously, I
had been using a fairly generic macro, ID to handle device names.  But I
was also using ID to handle all manner of strings, including function
names and parameter names.  The parameter use case could not allow '+'
and '-' to be allowed characters, otherwise it would interpret (for
example) "x-y" to be a single parameter rather than a subtraction of y
from x.

So, the new macro, DEV, is identical to the old macro ID, but it allows those
two extra characters.  It doesn't (yet) allow all the weird characters allowed
in NODE.  I don't yet know if this is needed or not, but the test
INVALID_CHARS/valid_char_expressions.cir only tests '+' and '-' in device
names.  So, for now, DEV doesn't include things like square brackets,
greater than, less than, '%', '^', etc.

DEV is used in the lexer whenever the context changes to a device-name
specific context.  ID is used otherwise for random strings.

I have also created unit tests in parserUnitTest.C and
complexParserUnitTest.C that checks that all these new characters work.