Overdue updates to complexParserUnitTests
A while ago I changed the group structure so that current solution
variables and voltage solution variables were handled separately.
Previously, they were treated as being the same kind of thing,
accessible by the same kind of function.  This proved to be untennable,
which is something I should have anticipated, given that currents can
also be lead currents, etc.

Back when I made the change, I updated the double-precision unit tests
(parserUnitTest.C) but not the complex version.  This commit corrects
that oversight.