Add proper support for DB (vdb, idb, ndb)
I had written code to parse VDB,IDB and NDB, but had not correctly set
the operator(s) up.  Also, as it turns out I made a mistake in the
parser file so encountering VDB was causing the code to seg fault.

The operators for VDB, IDB and NDB are now set up correctly, as an
expression tree that evaluates to 20.0*std::log10(std::abs(arg)).
Since I am building from existing expressions, which all have derivative
functions, the derivative support came automatically.

The seg faults for this are gone now, and it produces the correct
answer.