Beginnings of supporting phase in degrees
In Xyce, phase can be output in either degrees or radians.  Radians is
the default.   This commit starts to add code for supporting both.  I
had just supported radians in the new expression library before now.

This isn't done yet, but I wanted to commit it before I forgot about it,
as I have a number of other new expression developments going
on right now.