Fix so that function names can include # @, etc
This commit is similar to previous commits that applied to voltage node
names, to allow strange characters to be used in node names.  In this
case a handful of strange characters also needed to be allowed for
function names, including #, @, _ and the backtick character.

The # character took a tiny bit of extra work because I had treated it
as the beginning of a comment.  That was a holdover from last fall, when
I wasn't concerned with being 100% compliant with Xyce.  And, in
practice I had no tests that were using # as a comment delimeter. So,
that part of the lexer has been removed. (actually commented out at the
moment, but is slated for removal).

The changes in this commit allow the INVALID_CHARS/vald_chars_func_names
test to pass.