Complete implementation for gmin and phase units
For both the special variable GMIN and also the units of phase output, I
had started implementing them, but had not finished.  This commit
finishes both of them.  Most of the fixes were book-keeping (in the case
of phase) and also adding functions to the two group classes,
outputsXyceExpressionGroup and mainXyceExpressionGroup.

In a previous commit I mistakenly said that the default output units for
phase was radians.  That was incorrect, it got changed in 7.0 to
degrees.  The new expression library will now use the correct default,
and it also gets the user-specified netlist boolean for this setting.