Fix for memory mistakes, in the noise ops and tests
I was getting intermittent seg faults when I ran the complexParserUnitTest
program.  The main problem was a silly mistake (which I subsequently copied
to a few places) in the "noise" group that I created for the DNO and DNI
noise tests.  Fortunately, that meant the main problem was in the unit
test code, and not the expression library itself.

However, there were also 2 memory leaks in the new expression parser, also
having to do with DNO and DNI operators.  These mistakes were catastrophic -
just some memory that wasn't getting freed.  But they needed to be fixed.

With this commit, the complexParserUnitTest program runs thru valgrind
with zero errors again.