Fix CMake cURL library inclusion
CMake was finding cURL, but doing nothing with it, explicitly. If the
cURL libraries were in a standard location, it was not a problem;
however, that seems to not be the case on Windows. Now it should be
properly telling Xyce where the library is for linking.

Xyce/Xyce#2
Xyce/Xyce#1