Fix error in variable name
We were adding "FLEX_INCLUDE_DIR" to the list of include directories,
when the variable actually created by FindFLEX (and consequently
"find_package(FLEX)") is "FLEX_INCLUDE_DIRS"

This is unimportant on 90% of the systems where we build, where flex
include directories are basically "/usr/include" and so are
automaticlaly found, but it does matter on some where multiple
versions of Flex may be installed (e.g. an ancient version with the
core system and a more recent version from a package manager).