Try to force CMake not to use CURLConfig.cmake
As built on Windows, cURL has provided a CURLConfig.cmake that is
supposed to be getting used by CMake to set various configuration
variables.  This feature is ONLY getting used in recent versions of
CMake, but we're using one of those.

The documentation at
https://cmake.org/cmake/help/latest/module/FindCURL.html says taht this
can be disabled with a CURL_NO_CURL_CMAKE variable.  Let's try setting
it.

Xyce/Xyce#2