Try to force CURL cmake to work
When built with CMake, CURL provides it's own cmake config, which does
NOT do the same thing that FindCURL would do --- specifically, it does
NOT define CURL_LIBRARIES or CURL_INCLUDE_DIRS, instead it creates an
"import" library called CURL::libcurl.  This brings with it (in theory)
all the required other libraries that curl needs.

This commit *tries* to make use of that properly.  It may be completely
wrong and might not work at all.