Finish implementation of RF params in new expr
This doesn't *quite* complete it, but it completes it enough to pass the
two regression tests that were previously failing:

    Output/SPARAMS/sparams-ts2-dataFormat.cir
    MEASURE_AC/RFParamsOps.cir

These two tests now pass.

As I noted in my previous commit, they were failing b/c the function
AC::setRFParamsRequested was not being called early enough.  It needed
to be called prior to AC::doInit, and in my implementation it was not.
I have added some code to call this function (via the group) at the end
of the lexAndParse function.  This function is usually called early in
the Xyce parser, before any analysis is performed.

The one way in which this isn't *quite* a complete fix is that it won't
properly catch external dependencies in expressions.  So, if I have:

.global_param gp1 {SDB(1,1)}
.print AC {gp1}

Then I think it still won't be called early enough.  At the initial
parse of "gp1" it won't know that the external dependency depends on an
RF parameter.  So it thus won't call AC::setRFParamsRequested.

So, for now, that use case is broken.  But that use case isn't in the
test suite so far.