Optimization of tableOp operator
One of the goals of the new expression library has been for the new
library to be faster than the old one, especially for a handful of
problematic use cases.

One of those use cases is documented on the SON in bug 1176, which is
about table-based sources, when the table is really big.  This bug was
inspired by bug 1173, which was reported by Perry Robertson for the PWL
Vsrc.  He had been trying to produce a transient white noise source
using a PWL file consisting of 100,000 time,value pairs.  It turned out
that both the Vsrc and ABM/Bsrc implementation of PWL sources could not
handle this.

In spring of 2019, I fixed bug 1173 for the Vsrc.  The problem for the
Vsrc was all about management of breakpoints when there is a really
large number.    This mostly required careful application of
std::sort, std::unique, etc.

For the ABM/Bsrc use case, it also failed, but the failure was
unrelated to breakpoints.  The details are not very important, but they
appeared to be intractable, and are documented thoroughly in bug 1176.

Now that the new expression library is in pretty good shape, I decided
to dig out Perry's test case with the 100,000 PWL points.  Initially, it
ran  pretty slowly in the new expression library, but this was easy to
fix.  There were two issues causing the slow runtimes.

(1) Even though the breakpoints never changed, I was reporting all of
them back to the time integrator at every single time step.  All
100,000, over and over, meaning that the sorting and merging had to be
done every time on this relatively large data set.  This was very easy
to fix.

(2) The tableOp allows for entries in the table to be expressions and
not just numbers.  If they are non-constant expressions, they have to be
re-evaluated ever time, and then the table has to be re-initialized at
every val() call.    But, if you know that the entries in the table are
constant numbers, this re-evaluation should not be necessary, and the
initialization of the interpolator is not necessary either, except for
once in the constructor.    As it turns out, the logic I had in place to
check if the table was constant numbers was busted, so I was doing all
this extra work.

With this commit, issue (1) and issue (2) are fixed.  When I run
Perry's circuit now, using serial Xyce (but compiled with -O0, not -O3,
so not the fastest), the results are:

Vsrc version (circuit from bug 1173): Total Elapsed Run Time: 26.6071 seconds
Bsrc version (circuit from bug 1176): Total Elapsed Run Time: 20.6575 seconds

The second one (Bsrc version) is the one using the new expression
library.  So, gratifyingly, it is about 20% faster than the
N_DEV_SourceData PWL source class.    I would not have expected that, so
this warms my heart.