Add support for $write and $display verilog-a functions
$display is a synonym for $strobe.  $write is exactly like $strobe
except it doesn't write out a newline.

$write is used by HiSIM (all. over. the. place.).  And as long as I
was hacking in support for it, might as well do $display.

Tested only by regenerating all existing Verilog-A models and comparing to
previous --- none of those that use $strobe are impacted in any way,
so I didn't break that feature.

Changes to the three utils scripts are to correct the omission of one
of the non-free devices that I added to the code but forgot to add to
the script.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/26
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=846