Removes inclusion of unnecessary Epetra files / use of SDM
The AC header, PCE solver factory, and null time integration strategy
included Epetra headers but did not directly use Epetra objects.
These are being removed.

The measure package used an Epetra_SerialDenseVector object when it
could have used a Teuchos::SerialDenseVector object.  That was just
a simple swap of classes.

This is part of issue #25.