Removes unused Epetra headers and direct use of EpetraExt MatrixMarket IO
Some more unused Epetra headers were removed from the noise analysis class.

There was also a unnecessary inclusion of MatrixMarket IO headers and calls
directly to EpetraExt to write linear algebra objects to MatrixMarket files.
This can be cleaned up more, but this is a first pass.

The writeToFile method for multi-vectors has been extended to output
the MatrixMarket files format.

This is relevant to issue #25.