Make better search for Intel MLK FFT capability
I tried to make the MKL FFT search be a little more "standard". It's not
necessarily the best approach; but, at the very least, it will work for now,
and is a little cleaner than what was being done before.

I also changed some of the messages slightly, and did some comment cleaning.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1