Comment out debug outputs
Unfortunatley, I didn't catch these earlier.  In my earlier work (a
month or more ago) I was doing a lot of debugging of how to attach
.funcs to expressions.  During that work I added a lot of debug print
statements, but didn't put them inside of #ifdefs, and so it was easy to
forget about them.  Anyway, I re-discovered them today and have removed
them.