Improve ddt operator in several ways
I previously implemented the ddt operator to simply compute a
backward-euler finite difference, locally inside the operator class.

This was the equivalent of what the old library did when an expression
using DDT was invoked in the .PRINT line.  But it was not equivalent to
its usage in a Bsrc.  The old expression library, inside a Bsrc,  uses
the time integrator to compute time derivatives, and this allows them to
be the same order and integration method as the rest of the code.  So,
not just Backward Euler.

Interestingly, my BE-only implementation of ddt passed regression tests
using ddt on the .print line, and broke for some of the tests where ddt
was used in a bsrc, such as ABM/ddt.cir.

This commit re-implements a few functions that I previously had removed,
which allow the calling code to push into values for time derivatives,
and implements them (mostly) properly in the new expression class.  It
also un-comments the use of these functions in the Bsrc.

Finally, I also have reverted to doing better book-keeping for time
steps, which should (hopefully) fix some issues in ddt and sdt.   I had
futilely been trying to determine when to update time points in ddt and
sdt by inference, from things like the time step number, value of the
time, etc.  This approach is easy to get wrong, and I've decided to
abandon it.  So, now, the calling code has to call a "push" function,
i.e. "processSuccessfulTimeStep".

Unforunately, this new work on ddt (and this also applies to sdt) will
not quite be sufficient.  Both ddt and sdt are currently busted in the
new expression library, if they invoked inside of a .func that is called
by multiple expressions.  The reason is that their is state history
associated with both ddt and sdt.   But, if ddt is inside of a .func and
called by multiple expressions. that state history is not properly
managed, since it doesn't come into the ddtOp class via a function
argument to the constructor.

But what I am adding in this commit is code that does allow (I think)
sdt and ddt to work correctly as long as they are not inside a .func.